feat/authorization-finished

+++ src/modules/addresses/addresses.controller.ts
@@ -8,6 +8,8 @@ import {
   Delete,
   Request,
   UseGuards,
+  ForbiddenException,
+  HttpCode,
 } from '@nestjs/common';
 import { AddressesService } from './addresses.service';
 import { CreateAddressDto } from './dto/create-address.dto';
@@ -20,32 +22,68 @@ export class AddressesController {
 
   @UseGuards(JwtAuthGuard)
   @Post()
-  create(@Body() createAddressDto: CreateAddressDto, @Request() req) {
-    return this.addressesService.create(createAddressDto, req.user.id);
+  async create(@Body() createAddressDto: CreateAddressDto, @Request() req) {
+    const createdAddress = await this.addressesService.create(
+      createAddressDto,
+      req.user.id,
+    );
+    return createdAddress;
   }
 
   @UseGuards(JwtAuthGuard)
   @Get()
   findAll(@Request() req) {
     const { page, limit } = req.query;
-    return this.addressesService.findAll(page, limit);
+    if (req.user.admin) {
+      return this.addressesService.findAll(page, limit);
+    } else {
+      throw new ForbiddenException(
+        'Você não tem permissão para acessar esta rota',
+      );
+    }
   }
 
   @UseGuards(JwtAuthGuard)
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.addressesService.findOne(id);
+  async findOne(@Param('id') id: string, @Request() req) {
+    const oneAdress = await this.addressesService.findOne(id);
+    if (req.user.admin == true || req.user.id == oneAdress.personId) {
+      return oneAdress;
+    } else {
+      throw new ForbiddenException(
+        'Você não tem permissão para acessar esta rota',
+      );
+    }
   }
 
   @UseGuards(JwtAuthGuard)
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateAddressDto: UpdateAddressDto) {
-    return this.addressesService.update(id, updateAddressDto);
+  async update(
+    @Param('id') id: string,
+    @Body() updateAddressDto: UpdateAddressDto,
+    @Request() req,
+  ) {
+    const oneAdress = await this.addressesService.findOne(id);
+    if (req.user.admin == true || req.user.id === oneAdress.personId) {
+      return this.addressesService.update(id, updateAddressDto);
+    } else {
+      throw new ForbiddenException(
+        'Você não tem permissão para acessar esta rota',
+      );
+    }
   }
 
+  @HttpCode(204)
   @UseGuards(JwtAuthGuard)
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.addressesService.remove(id);
+  async remove(@Param('id') id: string, @Request() req) {
+    const oneAdress = await this.addressesService.findOne(id);
+    if (req.user.admin == true || req.user.id === oneAdress.personId) {
+      return this.addressesService.remove(id);
+    } else {
+      throw new ForbiddenException(
+        'Você não tem permissão para acessar esta rota',
+      );
+    }
   }
 }
